:recycle: api: moved CodeSession to src/certification/session/domain/services/
