+++ api/scripts/data-generation/generate-certif-cli.js
@@ -12,7 +12,7 @@ import lodash from 'lodash';
 
 const { maxBy } = lodash;
 import { logger } from '../../lib/infrastructure/logger.js';
-import { getNewSessionCode } from '../../lib/domain/services/session-code-service.js';
+import { getNewSessionCode } from '../../src/certification/session/domain/services/session-code-service.js';
 import { temporaryStorage } from '../../lib/infrastructure/temporary-storage/index.js';
 import {
   makeUserPixCertifiable,
